refactor: split uploadToS3()
